rename persons to people, fix react-router-dom import
