+++ src/features/people/PeopleDetails/index.js
